# Hash Join Vectorized collision checking