--- a/datafusion/core/src/physical_plan/joins/hash_join.rs
+++ b/datafusion/core/src/physical_plan/joins/hash_join.rs
@@ -26,7 +26,7 @@ use arrow::array::{
     UInt8Array,
 };
 use arrow::buffer::BooleanBuffer;
-use arrow::compute::{and, eq_dyn, filter, is_null, or, take};
+use arrow::compute::{and, eq_dyn, filter, is_null, or_kleene, take};
 use arrow::datatypes::{ArrowNativeType, DataType};
 use arrow::datatypes::{Schema, SchemaRef};
 use arrow::record_batch::RecordBatch;
@@ -1089,7 +1089,7 @@ fn eq_dyn_null(
             let left_is_null = is_null(left)?;
             let right_is_null = is_null(right)?;
 
-            or(&eq, &and(&left_is_null, &right_is_null)?)
+            or_kleene(&and(&left_is_null, &right_is_null)?, &eq)
         }
         _ => eq_dyn(left, right),
     }