# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #50

# Rationale for this change

Speed up for hash joins:

```

┏━━━━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┓
┃ Query        ┃     main ┃ vectorized_collision ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ QQuery 1     │ 186.15ms │             185.23ms │     no change │
│ QQuery 2     │  68.78ms │              60.46ms │ +1.14x faster │
│ QQuery 3     │  53.50ms │              49.64ms │ +1.08x faster │
│ QQuery 4     │  38.11ms │              38.57ms │     no change │
│ QQuery 5     │ 112.94ms │             101.61ms │ +1.11x faster │
│ QQuery 6     │  11.18ms │              10.97ms │     no change │
│ QQuery 7     │ 228.21ms │             197.74ms │ +1.15x faster │
│ QQuery 8     │  80.80ms │              75.27ms │ +1.07x faster │
│ QQuery 9     │ 166.03ms │             145.26ms │ +1.14x faster │
│ QQuery 10    │ 106.34ms │             100.25ms │ +1.06x faster │
│ QQuery 11    │  55.83ms │              49.45ms │ +1.13x faster │
│ QQuery 12    │  70.98ms │              69.73ms │     no change │
│ QQuery 13    │ 217.69ms │             192.88ms │ +1.13x faster │
│ QQuery 14    │  13.77ms │              12.82ms │ +1.07x faster │
│ QQuery 15    │  23.83ms │              23.92ms │     no change │
│ QQuery 16    │  52.41ms │              53.94ms │     no change │
│ QQuery 17    │ 699.16ms │             699.39ms │     no change │
│ QQuery 18    │ 621.07ms │             547.10ms │ +1.14x faster │
│ QQuery 19    │  61.40ms │              60.98ms │     no change │
│ QQuery 20    │ 218.51ms │             212.73ms │     no change │
│ QQuery 21    │ 341.74ms │             275.30ms │ +1.24x faster │
│ QQuery 22    │  34.32ms │              29.75ms │ +1.15x faster │
└──────────────┴──────────┴──────────────────────┴───────────────┘
```

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

# What changes are included in this PR?
 Changes the collision checking to be vectorized.

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

# Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->