@metesynnada 2023-06-20
There is a significant improvement in the equality check, nice work. Maybe then the `apply_join_filter_to_indices` can be addressed, since the intermediate batch can be huge in low cardinality cases 😃