@Dandandan 2023-06-21
Btw I think this can be applied to the symmetric hash join as well without changes.